+++ include/cover.h
@@ -18,7 +18,7 @@ void cover_init(cover_s *cover);
 
 void cover_make_patch_aligned_to_line(
     cover_s *cover, point_s row_data[NUM_LAYERS][MAX_POINTS_PER_LAYER],
-    int num_points[NUM_LAYERS], double apexZ0, double z_top, bool leftRight);
+    int num_points[NUM_LAYERS], float apexZ0, float z_top, bool leftRight);
 
 void cover_make_patch_shadow_quilt_from_edges(
     cover_s *cover, point_s row_data[NUM_LAYERS][MAX_POINTS_PER_LAYER],
