+++ src/point.cxx
@@ -1,7 +1,7 @@
 #include "point.h"
 #include <iostream>
 
-point_s point_init(int layer_num, double radius, double phi, double z) {
+point_s point_init(int layer_num, float radius, float phi, float z) {
   point_s p;
   p.layer_num = layer_num;
   p.radius = radius;
